SD/SD_Expt4_Sharding/requirements.txt
fastapi
uvicorn
mysql-connector-python
pymongo
jinja2
python-multipart
pandas